Fix: Change DeleteStarPlayerInPositionComman Ids to long

+++ BaseballStats.Application/Features/StarPlayerInPosition/Delete/DeleteStarPlayerInPositionCommand.cs
@@ -5,8 +5,8 @@ namespace BaseballStats.Application.Features.StarPlayerInPosition.Delete;
 
 public class DeleteStarPlayerInPositionCommand : ICommand<StarPlayerInPositionDto>
 {
-    public int PlayerId { get; set; }
-    public int SeriesId { get; set; }
-    public int SeasonId { get; set; }
+    public long PlayerId { get; set; }
+    public long SeriesId { get; set; }
+    public long SeasonId { get; set; }
     public string Position { get; set; } = null!;
 }
